Fixed rebase issues
