+++ VictoryCenter/VictoryCenter.DAL/Repositories/Realizations/Base/RepositoryWrapper.cs
@@ -1,10 +1,9 @@
 using System.Transactions;
 using VictoryCenter.DAL.Data;
 using VictoryCenter.DAL.Repositories.Interfaces.Base;
-using VictoryCenter.DAL.Repositories.Interfaces.TeamMembers;
 using VictoryCenter.DAL.Repositories.Interfaces.Categories;
+using VictoryCenter.DAL.Repositories.Interfaces.TeamMembers;
 using VictoryCenter.DAL.Repositories.Interfaces.Test;
-
 using VictoryCenter.DAL.Repositories.Realizations.Categories;
 using VictoryCenter.DAL.Repositories.Realizations.TeamMembers;
 using VictoryCenter.DAL.Repositories.Realizations.Test;
@@ -19,7 +18,6 @@ public class RepositoryWrapper : IRepositoryWrapper
     private ICategoriesRepository? _categoriesRepository;
     private ITeamMembersRepository? _teamMembersRepository;
 
-    private ITeamMemberRepository? _teamMemberRepository;
     public RepositoryWrapper(VictoryCenterDbContext context)
     {
         _victoryCenterDbContext = context;
@@ -27,7 +25,7 @@ public class RepositoryWrapper : IRepositoryWrapper
 
     public ITestRepository TestRepository => _testRepository ??= new TestRepository(_victoryCenterDbContext);
     public ICategoriesRepository CategoriesRepository => _categoriesRepository ??= new CategoriesRepository(_victoryCenterDbContext);
-    public ITeamMemberRepository TeamMemberRepository => _teamMemberRepository ??= new TeamMemberRepository(_victoryCenterDbContext);
+    public ITeamMembersRepository TeamMembersRepository => _teamMembersRepository ??= new TeamMembersRepository(_victoryCenterDbContext);
 
     public int SaveChanges()
     {
