+++ VictoryCenter/VictoryCenter.WebAPI/Controllers/TeamMembers/TeamMembersController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using VictoryCenter.BLL.Commands.TeamMembers.CreateTeamMember;
+using VictoryCenter.BLL.Commands.TeamMembers.Delete;
+using VictoryCenter.BLL.Commands.TeamMembers.Update;
 using VictoryCenter.BLL.DTOs.TeamMembers;
 using VictoryCenter.BLL.Queries.TeamMembers.GetByFilters;
 using VictoryCenter.BLL.Queries.TeamMembers.GetById;
@@ -26,5 +28,17 @@ public class TeamMembersController : BaseApiController
     public async Task<IActionResult> CreateTeamMember([FromBody] CreateTeamMemberDto createTeamMemberDto)
     {
         return HandleResult(await Mediator.Send(new CreateTeamMemberCommand(createTeamMemberDto)));
+    }
+
+    [HttpDelete("{id:long}")]
+    public async Task<IActionResult> DeleteTeamMember(long id)
+    {
+        return HandleResult(await Mediator.Send(new DeleteTeamMemberCommand(id)));
+    }
+
+    [HttpPut]
+    public async Task<IActionResult> UpdateTeamMember([FromBody] UpdateTeamMemberDto updateTeamMemberDto)
+    {
+        return HandleResult(await Mediator.Send(new UpdateTeamMemberCommand(updateTeamMemberDto)));
     }
 }
